+++ package.json
@@ -36,6 +36,7 @@
     "gulp-sourcemaps": "^2.6.4",
     "gulp-typedoc": "^2.2.0",
     "gulp-typescript": "^5.0.0-alpha.3",
+    "ipfs": "^0.32.3",
     "jest": "^23.5.0",
     "ts-jest": "^23.1.3",
     "tsconfig-paths": "^3.5.0",
